refactor: move prefetch git->benchmark (Phase 2)
populate_git_cache walks a BenchmarkExecution to collect repos, so it
belongs in the benchmark layer, not git. Moving it removes the last
git->benchmark import (it was TYPE_CHECKING-only) and leaves git
depending on nothing upward.

  git/prefetch.py -> benchmark/prefetch.py
  export moves from git/__init__ to benchmark/__init__
  caller cli/run.py + test patch targets updated

Cycle edges: only archive<->remote remains (Phase 3).
