+++ omnibenchmark/benchmark/__init__.py
@@ -2,6 +2,13 @@ from .benchmark import BenchmarkExecution
 from .benchmark import BenchmarkExecution as Benchmark  # Compatibility alias
 from ._node import BenchmarkNode
 from .validator import Validator
+from .prefetch import populate_git_cache
 
 
-__all__ = ["BenchmarkExecution", "Benchmark", "BenchmarkNode", "Validator"]
+__all__ = [
+    "BenchmarkExecution",
+    "Benchmark",
+    "BenchmarkNode",
+    "Validator",
+    "populate_git_cache",
+]
